Task_2/Level 2/requirements.txt
requests
beautifulsoup4
python-nmap
shodan
